# Improve performance and generalize supported block sizes