--- a/src/lib.rs
+++ b/src/lib.rs
@@ -179,52 +179,22 @@ trait Beamformer: Plugin {
     fn update_params(&mut self, ports: &mut Ports);
 }
 
-impl Plugin for Triforce {
-    type Ports = Ports;
+impl Triforce {
 
-    type InitFeatures = ();
-    type AudioFeatures = ();
-
-    fn new(info: &PluginInfo, _features: &mut ()) -> Option<Self> {
-        Some(Self {
-            hangle_curr: 0f32,
-            vangle_curr: 0f32,
-            freq_curr: 1000f32,
-            samples_since_last_update: u32::max_value(),
-            sample_rate: info.sample_rate() as f32,
-            covar_window: vec![
-                vec![Complex::new(0f32, 0f32); 256],
-                vec![Complex::new(0f32, 0f32); 256],
-                vec![Complex::new(0f32, 0f32); 256],
-            ],
-            array_geom: [ElemDistance { x: 0f32, y: 0f32 }; 3],
-            steering_vector: steering_vec(
-                90f32.to_radians(),
-                45f32.to_radians(),
-                1000f32,
-                [ElemDistance { x: 0f32, y: 0f32 }; 3],
-            ),
-            covar: DMatrix::zeros(3, 3),
-        })
-    }
-
-    fn run(&mut self, ports: &mut Ports, _features: &mut (), _: u32) {
-        Beamformer::update_params(self, ports);
+    fn process_slice(&mut self, mic1: &[f32], mic2: &[f32], mic3: &[f32], output: &mut [f32], t_win: f32) {
 
         // Steering vector is relative to Left/Top mic
         let inputs = vec![
-            analytic_signal(*ports.in_1),
-            analytic_signal(*ports.in_2),
-            analytic_signal(*ports.in_3),
+            analytic_signal(mic1),
+            analytic_signal(mic2),
+            analytic_signal(mic3),
         ];
+
         let num_samples = inputs[0].len();
-        if num_samples < 1024 {
-            return;
-        }
 
         // Update the covariance matrix. We use an overlapping window to smooth over
         // the transitions.
-        if self.samples_since_last_update as f32 >= (*ports.t_win / 1000.0) * self.sample_rate {
+        if self.samples_since_last_update as f32 >= (t_win / 1000.0) * self.sample_rate {
             self.samples_since_last_update = 0;
             self.covar_window[0].extend_from_slice(&inputs[0][0..768]);
             self.covar_window[1].extend_from_slice(&inputs[1][0..768]);
@@ -258,13 +228,53 @@ impl Plugin for Triforce {
         let re: Vec<f32> = out.iter().map(|z| z.re).collect();
 
         // Do all of our NFP and clamping here
-        for (real, output) in Iterator::zip(re.iter(), ports.out.iter_mut()) {
+        for (real, output) in Iterator::zip(re.iter(), output.iter_mut()) {
             if real.is_finite() && !real.is_nan() {
                 *output = real.clamp(-10f32, 10f32);
             } else {
                 *output = 0f32;
             }
         }
+
+    }
+}
+
+
+impl Plugin for Triforce {
+    type Ports = Ports;
+
+    type InitFeatures = ();
+    type AudioFeatures = ();
+
+    fn new(info: &PluginInfo, _features: &mut ()) -> Option<Self> {
+        Some(Self {
+            hangle_curr: 0f32,
+            vangle_curr: 0f32,
+            freq_curr: 1000f32,
+            samples_since_last_update: u32::max_value(),
+            sample_rate: info.sample_rate() as f32,
+            covar_window: vec![
+                vec![Complex::new(0f32, 0f32); 256],
+                vec![Complex::new(0f32, 0f32); 256],
+                vec![Complex::new(0f32, 0f32); 256],
+            ],
+            array_geom: [ElemDistance { x: 0f32, y: 0f32 }; 3],
+            steering_vector: steering_vec(
+                90f32.to_radians(),
+                45f32.to_radians(),
+                1000f32,
+                [ElemDistance { x: 0f32, y: 0f32 }; 3],
+            ),
+            covar: DMatrix::zeros(3, 3),
+        })
+    }
+
+    fn run(&mut self, ports: &mut Ports, _features: &mut (), samples: u32) {
+        Beamformer::update_params(self, ports);
+        if samples < 1024 {
+            return;
+        }
+        self.process_slice(&ports.in_1, &ports.in_2, &ports.in_3, &mut ports.out, *ports.t_win);
     }
 }