--- a/src/lib.rs
+++ b/src/lib.rs
@@ -174,6 +174,7 @@ pub struct Triforce {
     covar: DMatrix<Complex<f32>>,
     array_geom: [ElemDistance; 3],
     planner: Mutex<FftPlanner<f32>>,
+    weights: DVector<Complex<f32>>,
 }
 
 trait Beamformer: Plugin {
@@ -203,6 +204,7 @@ impl Triforce {
             ),
             covar: DMatrix::zeros(3, 3),
             planner: Mutex::new(FftPlanner::new()),
+            weights: DVector::zeros(3),
         }
     }
 
@@ -231,13 +233,13 @@ impl Triforce {
             self.covar_window[0] = inputs[0][768..1024].to_vec();
             self.covar_window[1] = inputs[1][768..1024].to_vec();
             self.covar_window[2] = inputs[2][768..1024].to_vec();
+            self.weights = mvdr_weights(&self.covar, &self.steering_vector);
         }
         else {
             self.samples_since_last_update += num_samples as u32;
         }
 
         // Get the MVDR weights
-        let w = mvdr_weights(&self.covar, &self.steering_vector);
 
         // Now we can finally do the beamforming
         let mut out = vec![Complex::zero(); num_samples];
@@ -249,7 +251,7 @@ impl Triforce {
             );
 
             // Conjugate-linear dot product
-            out[t] = w.dotc(&discrete);
+            out[t] = self.weights.dotc(&discrete);
         }
 
         // Now we need to revert the Hilbert transform and output the signal