--- a/src/lib.rs
+++ b/src/lib.rs
@@ -11,7 +11,7 @@
 
 use std::f32::consts::PI;
 use lv2::prelude::*;
-use nalgebra::{linalg::SVD, ComplexField, DMatrix, DVector, Vector3};
+use nalgebra::{linalg::SVD, ComplexField, Matrix3, Vector3};
 use rustfft::{num_complex::Complex, num_traits::Zero, FftPlanner};
 use std::sync::Mutex;
 
@@ -63,7 +63,7 @@ fn analytic_signal(planner: &mut FftPlanner<f32>, signal: &[f32]) -> Vec<Complex
 /// The steering vector is a representation of the phase delays at each microphone.
 /// It is calculated by taking the dot product of the array geometry matrix and the
 /// unit vector of the direction of arrival.
-fn steering_vec(theta: f32, phi: f32, f: f32, elems: [ElemDistance; 3]) -> DVector<Complex<f32>> {
+fn steering_vec(theta: f32, phi: f32, f: f32, elems: [ElemDistance; 3]) -> Vector3<Complex<f32>> {
     // Mic positions are relative to Left/Top to preserve x/y axis semantics
     let mic_positions: Vec<Vector3<f32>> =
         elems.iter().map(|e| Vector3::new(e.x, e.y, 0f32)).collect();
@@ -76,7 +76,7 @@ fn steering_vec(theta: f32, phi: f32, f: f32, elems: [ElemDistance; 3]) -> DVect
 
     // Calculate the steering vector by taking the array geometry, speed of sound,
     // and DOA unit vector
-    let mut steering_vector = DVector::from_element(mic_positions.len(), Complex::new(0f32, 0f32));
+    let mut steering_vector = Vector3::from_element(Complex::new(0f32, 0f32));
 
     for (i, mic_pos) in mic_positions.iter().enumerate() {
         let delay = mic_pos.dot(&u_dir) / C;
@@ -89,22 +89,21 @@ fn steering_vec(theta: f32, phi: f32, f: f32, elems: [ElemDistance; 3]) -> DVect
 
 /// There's nothing special about this, it's just a covariance matrix. It is always
 /// square.
-fn covariance(signals: &Vec<Vec<Complex<f32>>>) -> DMatrix<Complex<f32>> {
-    let n_mics = signals.len();
+fn covariance(signals: &[Vec<Complex<f32>>; 3]) -> Matrix3<Complex<f32>> {
     let n_samples = signals[0].len();
 
-    let mut covar = DMatrix::zeros(n_mics, n_mics);
+    let mut covar = Matrix3::zeros();
 
     for t in 0..n_samples {
-        let discrete: DVector<Complex<f32>> =
-            DVector::from_iterator(n_mics, signals.iter().map(|s| s[t]));
+        let discrete: Vector3<Complex<f32>> =
+            Vector3::new(signals[0][t], signals[1][t], signals[2][t]);
         covar += &discrete * discrete.adjoint();
     }
 
     // Our samples are shit, so we can't get a very nice covariance matrix.
     // Regularise the shit covariance matrix by introducing a constant value
     // across the identity
-    let reg = DMatrix::identity(covar.nrows(), covar.ncols())
+    let reg = Matrix3::identity()
         .map(|x: f32| Complex::new(x * 1e-4f32, 0f32));
 
     covar /= Complex::new(n_samples as f32, 0f32);
@@ -117,7 +116,7 @@ fn covariance(signals: &Vec<Vec<Complex<f32>>>) -> DMatrix<Complex<f32>> {
 /// w = (cov^-1 * sv) / (sv.adjoint() * (cov^-1 * sv)). Note that the denominator
 /// is the same as the conjugate-linear dot product of the steering vector and
 /// the numerator.
-fn mvdr_weights(cov: &DMatrix<Complex<f32>>, sv: &DVector<Complex<f32>>) -> DVector<Complex<f32>> {
+fn mvdr_weights(cov: &Matrix3<Complex<f32>>, sv: &Vector3<Complex<f32>>) -> Vector3<Complex<f32>> {
     // Since we have a numerically unstable covariance matrix, we can't take the
     // true inverse of it. Let's instead decompose it and take the pseudoinverse.
     let svd = SVD::new(cov.to_owned(), true, true);
@@ -169,12 +168,12 @@ pub struct Triforce {
     freq_curr: f32,
     sample_rate: f32,
     samples_since_last_update: u32,
-    covar_window: Vec<Vec<Complex<f32>>>,
-    steering_vector: DVector<Complex<f32>>,
-    covar: DMatrix<Complex<f32>>,
+    covar_window: [Vec<Complex<f32>>; 3],
+    steering_vector: Vector3<Complex<f32>>,
+    covar: Matrix3<Complex<f32>>,
     array_geom: [ElemDistance; 3],
     planner: Mutex<FftPlanner<f32>>,
-    weights: DVector<Complex<f32>>,
+    weights: Vector3<Complex<f32>>,
 }
 
 trait Beamformer: Plugin {
@@ -190,7 +189,7 @@ impl Triforce {
             freq_curr: 1000f32,
             samples_since_last_update: u32::max_value(),
             sample_rate,
-            covar_window: vec![
+            covar_window: [
                 vec![Complex::new(0f32, 0f32); 256],
                 vec![Complex::new(0f32, 0f32); 256],
                 vec![Complex::new(0f32, 0f32); 256],
@@ -202,9 +201,9 @@ impl Triforce {
                 1000f32,
                 [ElemDistance { x: 0f32, y: 0f32 }; 3],
             ),
-            covar: DMatrix::zeros(3, 3),
+            covar: Matrix3::zeros(),
             planner: Mutex::new(FftPlanner::new()),
-            weights: DVector::zeros(3),
+            weights: Vector3::zeros(),
         }
     }
 
@@ -245,10 +244,8 @@ impl Triforce {
         let mut out = vec![Complex::zero(); num_samples];
 
         for t in 0..num_samples {
-            let discrete: DVector<Complex<f32>> = DVector::from_iterator(
-                3, // number of mics
-                inputs.iter().map(|s| s[t]),
-            );
+            let discrete: Vector3<Complex<f32>> =
+                Vector3::new(inputs[0][t], inputs[1][t], inputs[2][t]);
 
             // Conjugate-linear dot product
             out[t] = self.weights.dotc(&discrete);