# Improving performance 

This PR starts by adding a `performance-test` target to the Makefile that processes 10 minutes of dummy contents and measure the time taken using `hyperfine`.

The next patches implement a bunch of optimization on top of that code (that should not change the output):

```
Initial code

  Time (mean ± σ):      2.139 s ±  0.029 s    [User: 2.133 s, System: 0.001 s]
  Range (min … max):    2.112 s …  2.182 s    10 runs
 
Cache fft plans

  Time (mean ± σ):     848.4 ms ±  30.1 ms    [User: 845.2 ms, System: 0.6 ms]
  Range (min … max):   811.8 ms … 877.9 ms    10 runs
 
Cache mvdr weights

  Time (mean ± σ):     664.7 ms ±  31.5 ms    [User: 661.9 ms, System: 0.8 ms]
  Range (min … max):   619.0 ms … 686.1 ms    10 runs
 
Use static vectors

  Time (mean ± σ):     495.2 ms ±  34.0 ms    [User: 492.7 ms, System: 1.0 ms]
  Range (min … max):   450.3 ms … 522.4 ms    10 runs
 
Use one less intermediate vector

  Time (mean ± σ):     490.4 ms ±  34.8 ms    [User: 488.5 ms, System: 0.5 ms]
  Range (min … max):   451.0 ms … 524.9 ms    10 runs
```

So processing is roughly 4x times faster after the optimizations.
Now 80% of the time is spent in the analytic transform and, more specifically, 65% in the FFT.

# Support other block sizes

The existing code was assuming blocks of 1024 samples (unfortunately, on my system, a fresh install of fedora 42 with gnome, the default block size was 256, so the microphone was broken).
Now it works with any size, while trying to compute the same coveriance window as if the blocks were 1024-sized:

- Larger blocks are split into chunks of size 1024 (except for the last chunk which has a size between 1024 and 2047); the result should be audibly identical
- Smaller blocks are accumulated until reaching an appropriately sized window. The result can still be different since the analytic transform is done blockwise on smaller blocks (it would be worth checking if an overlapping analytic transform improve the reconstruction).