@chadmed 2025-05-10
Merged manually with some minor refactoring as discussed on Matrix. Thanks heaps!